refactor(chat): decompose ChatHistoryDrawer into modular components
BREAKING DOWN COMPLEXITY: 243 lines → 121 lines (-50% reduction!)

## Changes:
- Extract chat history logic into custom hooks:
  • useChatHistory (conversation loading, CRUD operations)
  • useDrawerEffects (keyboard, scroll lock)
- Create isolated UI components:
  • ChatHistoryItem (conversation list item)
  • ChatHistoryEmpty (empty state)
  • ChatHistoryLoading (loading spinner)
- Main component now focuses on layout composition

## Benefits:
✅ Massive code reduction (243 → 121 lines, -50%)
✅ Better testability (isolated hooks and components)
✅ Single responsibility (each file has one job)
✅ Reusable components (ChatHistoryItem, useDrawerEffects)
✅ Cleaner imports and dependencies
✅ Easier to maintain and extend

## Files Changed:
- ChatHistoryDrawer.tsx (243 → 121 lines, -50%)
+ useChatHistory.ts (96 lines, business logic)
+ useDrawerEffects.ts (28 lines, a11y effects)
+ ChatHistoryItem.tsx (61 lines, list item UI)
+ ChatHistoryEmpty.tsx (7 lines, empty state)
+ ChatHistoryLoading.tsx (6 lines, loading UI)

Total: +198 new organized code, -122 old monolithic code
Net: +76 lines, but much better structure!

## Code Quality:
- Extracted callbacks with proper dependencies
- Separated effects from business logic
- Component composition over monolith
- Type-safe interfaces for all components
