+++ src/components/chat/components/ChatHistoryLoading.tsx
@@ -1,0 +1,7 @@
+export function ChatHistoryLoading() {
+  return (
+    <div className="flex justify-center py-8">
+      <div className="animate-spin rounded-full h-6 w-6 border-b-2 border-accent-primary"></div>
+    </div>
+  );
+}
